+++ cpp/util/NdHistogram.cc
@@ -6,11 +6,7 @@ NdHistogram::NdHistogram() : m_box(box::Box()), m_frame_counter(0), m_n_ref(0), 
 
 void NdHistogram::resetGeneral(unsigned int bin_size)
 {
-    for (tbb::enumerable_thread_specific<unsigned int*>::iterator i = m_local_bin_counts.begin();
-         i != m_local_bin_counts.end(); ++i)
-    {
-        memset((void*) (*i), 0, sizeof(unsigned int) * bin_size);
-    }
+    m_local_bin_counts.reset();
     this->m_frame_counter = 0;
     this->m_reduce = true;
 }