+++ tests/test_density_LocalDensity.py
@@ -1,9 +1,7 @@
 import numpy as np
 import numpy.testing as npt
 import freud
-import warnings
 import unittest
-import os
 
 
 class TestLD(unittest.TestCase):
@@ -16,6 +14,14 @@ class TestLD(unittest.TestCase):
         self.pos = np.array(np.random.random(size=(10000, 3)),
                             dtype=np.float32) * 10 - 5
         self.ld = freud.density.LocalDensity(3, 1, 1)
+
+        # Test access
+        with self.assertRaises(AttributeError):
+            self.ld.density
+        with self.assertRaises(AttributeError):
+            self.ld.num_neighbors
+        with self.assertRaises(AttributeError):
+            self.ld.box
 
     @unittest.skip('Skipping slow LocalDensity test.')
     def test_compute_api(self):
@@ -31,6 +37,12 @@ class TestLD(unittest.TestCase):
     def test_density(self):
         """Test that LocalDensity computes the correct density at each point"""
         self.ld.compute(self.box, self.pos, self.pos)
+
+        # Test access
+        self.ld.density
+        self.ld.num_neighbors
+        self.ld.box
+
         self.assertTrue(self.ld.box == freud.box.Box.cube(10))
 
         npt.assert_array_less(np.fabs(self.ld.density - 10.0), 1.5)