+++ ChangeLog.md
@@ -4,9 +4,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## next
 
+### Added
+* `AttributeError` is raised when one tries to access an attribute that has not yet been computed.
+
 ### Changed
 * Removed extra C++ includes to speed up builds.
 * The C++ style is now based on clang-format.
+* Refactored C++ handling of thread-local storage.
 
 ### Fixed
 * Calling setNumThreads works correctly even if a parallel compute method has already been called.